Fix terminal hotkey PR feedback
